Introduce JabKit (by moving much Gui launching code to there)
